Added heartbeat builder
